Create and switch sessions in tmux quickly
`tm mysession`

Whether or not you are in tmux, the tm function will switch to the given
session if it already exists. If it doesn't already exist, it will
create it and then switch to it.

I found the basics of this function somewhere on the internet. Apologies
to the original author, but I can't find it anymore.