Infer working directory if session name in CDPATH
Automatically set the working directory (the directory windows in a
session begin in) of new sessions to project folders, assuming that they
live in ~/code. The cdpath portion could be removed if cdpath were
already set to something other than `.`.

Using CDPATH necessitated switching the script to zsh instead of sh.